fix: soundess in verify_zero - bound q in 165 bits and val.d2 in 251 bits
after accounting for limbs being able to be 3*BASE our soundness checks were bound to 167bits causing LHS to be bigger than CAIRO prime.
r1 was inside 165 bits to begin with, for r2 we had to shift some bits from SECP_REM1 limb into SECP_REM2 allowing it to fit inside 165 bits.

Also fix up nondet_bigint3 to guarantee that d2 limb is guaranteed to be in [0, BASE) guaranteeing that val.d2 is in (-2**250, 2**250)